+++ src/lang/subprogram.cpp
@@ -92,26 +92,35 @@ int64_t Subprogram::search(const Program &main, const Program &sub,
   return main_pos - sub_pos;
 }
 
-void updateOperand(Operand &op, int64_t target, int64_t largest_used) {
+void updateOperand(Operand &op, int64_t start, int64_t shared_region_length,
+                   int64_t largest_used) {
   if (op.type != Operand::Type::DIRECT) {
     return;
   }
-  if (op.value.asInt() == Program::INPUT_CELL) {
-    op.value = target;
+  auto v = op.value.asInt();
+  if (v < shared_region_length) {
+    op.value += start;
   } else {
     op.value += largest_used;
   }
 }
 
-bool prepareEmbedding(int64_t id, Program &sub) {
+bool prepareEmbedding(int64_t id, Program &sub, Operation::Type embeddingType) {
   // load and check program to be embedded
+  std::string path;
+  if (embeddingType == Operation::Type::SEQ) {
+    path = ProgramUtil::getProgramPath(id);
+  } else if (embeddingType == Operation::Type::PRG) {
+    path = ProgramUtil::getProgramPath(id, "prg", "P");
+  } else {
+    throw std::runtime_error("Unsupported embedding type");
+  }
   Parser parser;
-  sub = parser.parse(ProgramUtil::getProgramPath(id));
+  sub = parser.parse(path);
   if (ProgramUtil::hasIndirectOperand(sub)) {
     return false;
   }
-  // prepare program for embedding
-  // remove nops and comments
+  // prepare program for embedding: remove nops and comments
   ProgramUtil::removeOps(sub, Operation::Type::NOP);
   for (auto &op : sub.ops) {
     if (op.type != Operation::Type::SEQ) {
@@ -120,7 +129,16 @@ bool prepareEmbedding(int64_t id, Program &sub) {
   }
   // find cells that are read and uninitialized
   std::set<int64_t> initialized, uninitialized;
-  initialized.insert(Program::INPUT_CELL);
+  if (embeddingType == Operation::Type::SEQ) {
+    initialized.insert(Program::INPUT_CELL);
+  } else if (embeddingType == Operation::Type::PRG) {
+    auto numInputs = sub.getDirective("inputs");
+    for (int64_t i = 0; i < numInputs; i++) {
+      initialized.insert(i);
+    }
+  } else {
+    throw std::runtime_error("Unsupported embedding type");
+  }
   ProgramUtil::getUsedUninitializedCells(sub, initialized, uninitialized);
   // initialize cells that are read and were uninitialized
   for (auto cell : uninitialized) {
@@ -136,34 +154,40 @@ bool Subprogram::unfold(Program &main) {
   if (ProgramUtil::hasIndirectOperand(main)) {
     return false;
   }
-  // TODO unfold prg instructions, too
-  // find first seq operation
-  int64_t seq_index = -1;
+  // find first seq or prg operation
+  int64_t pos = -1;
   for (size_t i = 0; i < main.ops.size(); i++) {
-    if (main.ops[i].type == Operation::Type::SEQ) {
-      seq_index = i;
+    if (main.ops[i].type == Operation::Type::SEQ ||
+        main.ops[i].type == Operation::Type::PRG) {
+      pos = i;
       break;
     }
   }
-  if (seq_index < 0) {
+  if (pos < 0) {
     return false;
   }
-  auto sub_id = main.ops[seq_index].source.value.asInt();
+  const auto &emb_op = main.ops[pos];
+  auto sub_id = emb_op.source.value.asInt();
   Program sub;
-  if (!prepareEmbedding(sub_id, sub)) {
+  if (!prepareEmbedding(sub_id, sub, emb_op.type)) {
     return false;
   }
   // shift used operands
-  auto target = main.ops[seq_index].target.value.asInt();
-  auto largest_used = ProgramUtil::getLargestDirectMemoryCell(main);
+  int64_t start = main.ops[pos].target.value.asInt();
+  int64_t shared_region_length = 1;
+  if (emb_op.type == Operation::Type::PRG) {
+    shared_region_length = std::max<int64_t>(sub.getDirective("inputs"),
+                                             sub.getDirective("outputs"));
+  }
+  int64_t largest_used = ProgramUtil::getLargestDirectMemoryCell(main);
   for (auto &op : sub.ops) {
-    updateOperand(op.target, target, largest_used);
-    updateOperand(op.source, target, largest_used);
+    updateOperand(op.target, start, shared_region_length, largest_used);
+    updateOperand(op.source, start, shared_region_length, largest_used);
   }
   // delete seq operation
-  main.ops.erase(main.ops.begin() + seq_index);
+  main.ops.erase(main.ops.begin() + pos);
   // embed program
-  main.ops.insert(main.ops.begin() + seq_index, sub.ops.begin(), sub.ops.end());
+  main.ops.insert(main.ops.begin() + pos, sub.ops.begin(), sub.ops.end());
   return true;
 }
 
@@ -189,7 +213,7 @@ bool Subprogram::autoUnfold(Program &main) {
 bool Subprogram::shouldFold(const Program &main) {
   int64_t level = 0;
   int64_t numLoops = 0;
-  bool hasRootSeq = false;
+  bool hasRootRef = false;
   for (const auto &op : main.ops) {
     switch (op.type) {
       case Operation::Type::LPB:
@@ -199,51 +223,54 @@ bool Subprogram::shouldFold(const Program &main) {
       case Operation::Type::LPE:
         level--;
         break;
+      case Operation::Type::PRG:
       case Operation::Type::SEQ:
         if (level == 0) {
-          hasRootSeq = true;
+          hasRootRef = true;
         }
         break;
       default:
         break;
     }
   }
-  return (numLoops > 1) || (numLoops > 0 && hasRootSeq);
+  return (numLoops > 1) || (numLoops > 0 && hasRootRef);
 }
 
 bool Subprogram::fold(Program &main, Program sub, size_t subId,
-                      std::map<int64_t, int64_t> &cell_map) {
+                      std::map<int64_t, int64_t> &cellMap) {
   // prepare and check subprogram
   ProgramUtil::removeOps(sub, Operation::Type::NOP);
   if (sub.ops.empty()) {
     return false;
   }
-  int64_t output_cell = Program::OUTPUT_CELL;
+  int64_t outputCell = Program::OUTPUT_CELL;
   auto last = sub.ops.back();
   if (sub.ops.size() > 1 && last.type == Operation::Type::MOV &&
       last.target == Operand(Operand::Type::DIRECT, Program::OUTPUT_CELL) &&
       last.source.type == Operand::Type::DIRECT) {
-    output_cell = last.source.value.asInt();
+    outputCell = last.source.value.asInt();
     sub.ops.pop_back();
   }
-  auto main_pos = Subprogram::search(main, sub, cell_map);
-  // no indirect ops allowed
+  auto mainPos = Subprogram::search(main, sub, cellMap);
+  // no indirect operands  allowed
   if (ProgramUtil::hasIndirectOperand(main) ||
       ProgramUtil::hasIndirectOperand(sub)) {
     return false;
   }
   // perform folding on main program
-  const Number mapped_input(cell_map.at(Program::INPUT_CELL));
-  const Number mapped_output(cell_map.at(output_cell));
-  main.ops.erase(main.ops.begin() + main_pos,
-                 main.ops.begin() + main_pos + sub.ops.size());
-  main.ops.insert(main.ops.begin() + main_pos,
+  const Number mappedInput(cellMap.at(Program::INPUT_CELL));
+  const Number mappedOutput(cellMap.at(outputCell));
+  main.ops.erase(main.ops.begin() + mainPos,
+                 main.ops.begin() + mainPos + sub.ops.size());
+  main.ops.insert(main.ops.begin() + mainPos,
                   Operation(Operation::Type::SEQ,
-                            Operand(Operand::Type::DIRECT, mapped_input),
+                            Operand(Operand::Type::DIRECT, mappedInput),
                             Operand(Operand::Type::CONSTANT, Number(subId))));
-  main.ops.insert(main.ops.begin() + main_pos + 1,
-                  Operation(Operation::Type::MOV,
-                            Operand(Operand::Type::DIRECT, mapped_output),
-                            Operand(Operand::Type::DIRECT, mapped_input)));
+  if (mappedOutput != mappedInput) {
+    main.ops.insert(main.ops.begin() + mainPos + 1,
+                    Operation(Operation::Type::MOV,
+                              Operand(Operand::Type::DIRECT, mappedOutput),
+                              Operand(Operand::Type::DIRECT, mappedInput)));
+  }
   return true;
 }